requirements.txt
numpy
pandas
matplotlib
opencv-python-headless
scikit-learn
streamlit
pillow
